fix: convert user id to int 64

+++ utils/jwt.go
@@ -58,7 +58,7 @@ func VerifyToken(token string) (int64, error) {
 	}
 
 	//email := claims["email"].(string)
-	userId := claims["userId"].(int64)
+	userId := int64(claims["userId"].(float64))
 
 	return userId, nil
 }
